cmd/go-markdown/spanelements/emoji.go
package spanelements

import (
	"github.com/pjotrscholtze/go-markdown/cmd/go-markdown/entity"
	"github.com/pjotrscholtze/go-markdown/cmd/go-markdown/util"
)

func parseLineEmojiElement(input []entity.MarkdownElement) []entity.MarkdownElement {
	res := make([]entity.MarkdownElement, 0)
	for _, entry := range input {
		if entry.Kind() != entity.ElementKindText {
			res = append(res, entry)
			continue
		}
		for _, entry := range util.FindPatternsAndNonPatterns(`:[A-Za-z0-9]+:`, entry.AsMarkdownString(), entity.ElementKindEmoji, entity.ElementKindText) {
			if entry.Type == entity.ElementKindEmoji {
				res = append(res, entity.NewEmojiMarkdownElement(entry.Content))
			} else {
				res = append(res, &entity.LineElement{
					Type:    entry.Type,
					Content: entry.Content,
				})
			}
		}
	}
	return res
}
